Move single-phase tests out of test_imp.
